requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.1
numpy==1.26.0
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
openai==1.2.0
pyyaml==6.0.1
pathlib==1.0.1
